Remove GitHub links
The point of this project: encourage people to hack code in public. The
idea is that if people announce that they are doing it, others might
join in. If one person does it, they're just crazy; two people, well
they're a little off. Three people? They may think it's an organization.
And 50 people tweeting with the `#codesthlm` hashtag, sitting together,
hacking on cool projects, designers and developers alike - well, that's
a movement.

Not the point of this project: arguing over the typeface.